requirements.txt
discord.py-self
aiohttp
python-dotenv
